Enhencement for improving test coverage
- Code separation rest handler & message creation
- Check code coverage for only message creation
- Skip coverage state in rest handler
