+++ scripts/tests_with_cover.sh
@@ -4,7 +4,7 @@ echo "" > coverage.txt
 for d in $(go list ./...); do
   go test -v -race -coverprofile=profile.out $d
   if [ -f profile.out ]; then
-    cat profile.out >> coverage.txt
+    cat profile.out | grep -v "client/cli" | grep -v "client/rest" | grep -v "friday/cmd" >> coverage.txt
     rm profile.out
   fi
 done
